Spit Assignmen recycler adapter
refs:
affects:
release note:

test plan:
